distribute schema and resolver file
